+++ Sources/ElasticSwift/Queries/Builders/FullTextBuilders.swift
@@ -14,10 +14,10 @@ public class MatchQueryBuilder: QueryBuilder {
     var field: String?
     var value: String?
     var isFuzzy: Bool?
-    var boost: Float?
+    var boost: Decimal?
     var `operator`: MatchQueryOperator?
     var zeroTermQuery: ZeroTermQuery?
-    var cutoffFrequency: Float?
+    var cutoffFrequency: Decimal?
     
     public var query: Query {
         get {
@@ -35,7 +35,7 @@ public class MatchQueryBuilder: QueryBuilder {
         return self
     }
     
-    public func set(cutoffFrequency: Float) -> Self {
+    public func set(cutoffFrequency: Decimal) -> Self {
         self.cutoffFrequency = cutoffFrequency
         return self
     }
@@ -55,7 +55,7 @@ public class MatchQueryBuilder: QueryBuilder {
         return self
     }
     
-    public func set(boost: Float) -> Self {
+    public func set(boost: Decimal) -> Self {
         self.boost = boost
         return self
     }
@@ -132,7 +132,7 @@ public class MultiMatchQueryBuilder: QueryBuilder {
     var value: String?
     var fields: [String]?
     var type: MultiMatchQueryType?
-    var tieBreaker: Float?
+    var tieBreaker: Decimal?
     
     public var query: Query {
         get {
@@ -140,7 +140,7 @@ public class MultiMatchQueryBuilder: QueryBuilder {
         }
     }
     
-    public func set(boost: Float) -> Self {
+    public func set(boost: Decimal) -> Self {
         return self
     }
     
@@ -155,7 +155,7 @@ public class MultiMatchQueryBuilder: QueryBuilder {
         return self
     }
     
-    public func set(tieBreaker: Float) -> Self {
+    public func set(tieBreaker: Decimal) -> Self {
         self.tieBreaker = tieBreaker
         return self
     }
@@ -167,7 +167,7 @@ public class MultiMatchQueryBuilder: QueryBuilder {
 public class CommonTermsQueryBuilder: QueryBuilder {
     
     var value: String?
-    var cutoffFrequency: Float?
+    var cutoffFrequency: Decimal?
     var lowFrequencyOperator: String?
     var highFrequencyOperator: String?
     var minimumShouldMatch: Int?
@@ -180,7 +180,7 @@ public class CommonTermsQueryBuilder: QueryBuilder {
         }
     }
     
-    public func set(boost: Float) -> Self {
+    public func set(boost: Decimal) -> Self {
         return self
     }
     
@@ -189,7 +189,7 @@ public class CommonTermsQueryBuilder: QueryBuilder {
         return self
     }
     
-    public func set(cutoffFrequency: Float) -> Self {
+    public func set(cutoffFrequency: Decimal) -> Self {
         self.cutoffFrequency = cutoffFrequency
         return self
     }
@@ -237,7 +237,7 @@ public class QueryStringQueryBuilder: QueryBuilder {
     var fuzzyPrefixLength: Int?
     var fuzzyTranspositions: Bool?
     var phraseSlop: Int?
-    var boost: Float?
+    var boost: Decimal?
     var autoGeneratePhraseQueries: Bool?
     var analyzeWildcard: Bool?
     var maxDeterminizedStates: Int?
@@ -347,7 +347,7 @@ public class QueryStringQueryBuilder: QueryBuilder {
         self.defaultOperator = defaultOperator
         return self
     }
-    public func set(boost: Float) -> Self {
+    public func set(boost: Decimal) -> Self {
         self.boost = boost
         return self
     }
